+++ src/addons/meta_human_dna/editors/rbf_editor/ui.py
@@ -20,10 +20,10 @@ class CHARACTER_DNA_UL_bone_selection(bpy.types.UIList):
         self,
         context: "Context",
         layout: bpy.types.UILayout,
-        data: "MetahumanWindowMangerProperties",
+        data: "CharacterWindowMangerProperties",
         item: "RBFDrivenBoneSelectionItem",
         icon: int | None,
-        active_data: "MetahumanWindowMangerProperties",
+        active_data: "CharacterWindowMangerProperties",
         active_prop_name: str,
     ) -> None:
         row = layout.row(align=True)